CountryByIp, Less Structured, Services Renamed, Build and js comments modified.
CountryByIp, Less Structured, Services Renamed, Build and js comments
modified.
- SQL script added for CountryByIp.
- CountryByIp added to the database. Service hasn't written yet.
- Less files organized , few still needs to be organized. Every pages
needs to have variables.
- Javascript comments added and still needs to be organized.
- Partial controller name is changed to services and renamed everywhere.
- Services controller name is coming from database and everything is
configured.

+++ DevBootstrapper/Models/POCO/IdentityCustomization/CountryDetectByIP.cs
@@ -5,8 +5,8 @@ namespace DevBootstrapper.Models.POCO.IdentityCustomization {
         [Key]
         public int CountryDetectByIPID { get; set; }
 
-        public int BeginingIP { get; set; }
-        public int EndingIP { get; set; }
+        public long BeginingIP { get; set; }
+        public long EndingIP { get; set; }
         public int CountryID { get; set; }
     }
 }
